Crate chipolata

Structs

ChipolataError

An Error struct used to bubble up Chipolata errors to the hosting application. This wraps the more specific ErrorDetail error enum, and provides overall processor state context at the point of the failure

Display

An abstraction of the CHIP-8 frame buffer.

Memory

An abstraction of the CHIP-8 memory space.

Options

A struct to allow specification of Chipolata start-up parameters.

Processor

An abstraction of the CHIP-8 processor, and the core public interface to the Chipolata crate.

Program

An abstraction of a CHIP-8 ROM, ready for loading into the Chipolata emulator.

Stack

An abstraction of the CHIP-8 stack, used for holding return addresses from function calls.

Enums

EmulationLevel

An enum to indicate which extension of CHIP-8 is to be emulated. See external documentation for details of the differences in each case.

ErrorDetail

An Error enum used throughout the Chipolata crate to communicate details of runtime errors that have occurred.

ProcessorStatus

An enum used internally within the Chipolata crate to keep track of the processor execution status.

StateSnapshot

An enum with variants representing the different Chipolata state snapshots that can be returned to hosting applications for processing

StateSnapshotVerbosity

An enum used to indicate which variant of StateSnapshot should be returned when a call is made to Processor::export_state_snapshot().

Constants

COSMAC_VIP_PROCESSOR_SPEED_HERTZ

The original COSMAC VIP processor speed in hertz. When instantiating an Options instance to pass to Chipolata, this value should normally be supplied as the starting Options::processor_speed_hertz choice when specifying EmulationLevel::Chip8 with variable_cycle_timing set to true. to be true.