Enum chipolata::ErrorDetail

pub enum ErrorDetail {
    UnknownInstruction {
        opcode: u16,
    },
    UnimplementedInstruction {
        opcode: u16,
    },
    OperandsOutOfBounds {
        operands: HashMap<String, usize>,
    },
    PopEmptyStack,
    PushFullStack,
    MemoryAddressOutOfBounds {
        address: u16,
    },
    InvalidKey {
        key: u8,
    },
    FileError {
        file_path: String,
    },
    StateTransitionError {
        old_state: ProcessorStatus,
        new_state: ProcessorStatus,
    },
    UnknownError,
}

An Error enum used throughout the Chipolata crate to communicate details of runtime errors that have occurred.

Instances of ErrorDetail are bubbled-up to the hosting application through the public API methods.

Variants

UnknownInstruction

Fields

opcode: u16

An unrecognised opcode was read from memory

UnimplementedInstruction

Fields

opcode: u16

A valid opcode was read from memory but which is not implemented by Chipolata

OperandsOutOfBounds

Fields

operands: HashMap<String, usize>

One or more operands fall outside expected ranges and cannot be safely used The HashMap field holds the name of each potential faulty operand and its value

PopEmptyStack

An attempt was made to pop an item off the Chipolata stack while it is empty

PushFullStack

An attempt was made to push an item on to the Chipolata stack while it is full

MemoryAddressOutOfBounds

Fields

address: u16

An attempt was made to read/write from an address outside the addressable range

InvalidKey

Fields

key: u8

A key ordinal was referenced that is outside the valid CHIP-8 keypad range (0x0 to 0xF)

FileError

Fields

file_path: String

Error used for any file I/O issues

StateTransitionError

Fields

old_state: ProcessorStatus

new_state: ProcessorStatus

Error causes by invalid processor state transition

UnknownError

General bucket for any unknown issues (to return something rather than panicking)