Enum chipolata::StateSnapshot

pub enum StateSnapshot {
    MinimalSnapshot {
        frame_buffer: Display,
        status: ProcessorStatus,
        processor_speed: u64,
        play_sound: bool,
        cycles: usize,
    },
    ExtendedSnapshot {

        frame_buffer: Display,
        status: ProcessorStatus,
        processor_speed: u64,
        play_sound: bool,
        cycles: usize,
        stack: Stack,
        memory: Memory,
        program_counter: u16,
        index_register: u16,
        variable_registers: [u8; 16],
        rpl_registers: [u8; 8],
        delay_timer: u8,
        sound_timer: u8,
        high_resolution_mode: bool,
        emulation_level: EmulationLevel,
    
},
}

An enum with variants representing the different Chipolata state snapshots that can be returned to hosting applications for processing

Variants

MinimalSnapshot

Fields

frame_buffer: Display

status: ProcessorStatus

processor_speed: u64

play_sound: bool

cycles: usize

Minimal snapshot containing only the frame buffer state, processor status, and a boolean to indicate whether a sound should be playing

ExtendedSnapshot

Fields

frame_buffer: Display

status: ProcessorStatus

processor_speed: u64

play_sound: bool

cycles: usize

stack: Stack

memory: Memory

program_counter: u16

index_register: u16

variable_registers: [u8; 16]

rpl_registers: [u8; 8]

delay_timer: u8

sound_timer: u8

high_resolution_mode: bool

emulation_level: EmulationLevel

Extended snapshot containing the minimal state along with all registers, stack and memory