Struct chipolata::Memory

pub struct Memory {
    pub bytes: [u8; 4096],
    /* private fields */
}

An abstraction of the CHIP-8 memory space.

Fields
bytes: [u8; 4096]

A stack-allocated array of bytes representing the entire CHIP-8 memory space

Implementations

impl Memory

pub fn read_byte(&self, address: usize) -> Result<u8, ErrorDetail>

Returns a copy of the byte in memory at the specified address. If the address is outside the addressable range, returns ErrorDetail::MemoryAddressOutOfBounds.

Arguments

address - the memory address at which the byte should be read

pub fn read_bytes(
&self,
start_address: usize,
num_bytes: usize
) -> Result<&[u8], ErrorDetail>

Returns an array slice from memory as per the specified start address and number of bytes. If the operands are such that the array slice would extend beyond addressable memory then returns ErrorDetail::MemoryAddressOutOfBounds.

Arguments

start_address - the memory address at the start of the range from which to read

num_bytes - the number of bytes to read from memory

pub fn read_two_bytes(&self, start_address: usize) -> Result<u16, ErrorDetail>

Returns a 16-bit unsigned integer constructed by reading two consecutive bytes from memory starting from the specified address. The construction is big-endian. In the unlikely event that the second byte would fall outside the addressable memory space, this returns ErrorDetail::MemoryAddressOutOfBounds.

The method is generally used as a convenience for reading opcodes from memory, as CHIP-8 opcodes are 16-bits in size.

Arguments

start_address - the memory address of the first (most significant) byte to read

pub fn max_addressable_size(&self) -> usize

Returns the size of the addressable memory space in bytes